Move tests out of tests/ folder for now.
