+++ trafficlight/sample_tests/video/ec_basic_example.py
