+++ trafficlight/sample_tests/video/handle_invite_links_registered_test.py
@@ -1,10 +1,10 @@
 from trafficlight.client_types import ElementCall
 from trafficlight.internals.client import ElementCallClient
 from trafficlight.internals.test import Test
-from trafficlight.tests.video.handle_invite_base import InviteLinksMixin
+from trafficlight.sample_tests.video.handle_invite_base import InviteLinksMixin
 
 
-class OneRegisteredInviteLinksTest(Test, InviteLinksMixin):
+class TwoRegisteredInviteLinksTest(Test, InviteLinksMixin):
     def __init__(self) -> None:
         super().__init__()
         self._client_under_test([ElementCall()], "alice")
@@ -12,6 +12,7 @@ class OneRegisteredInviteLinksTest(Test, InviteLinksMixin):
 
     async def run(self, alice: ElementCallClient, bob: ElementCallClient) -> None:
         await alice.register()
-        await bob.guest_user()
+        await bob.register()
         await alice.set_display_name()
+        await bob.set_display_name()
         await self._run_test(alice, bob)
