+++ trafficlight/sample_tests/video/handle_invite_links_two_guests_test.py
@@ -1,18 +1,16 @@
 from trafficlight.client_types import ElementCall
 from trafficlight.internals.client import ElementCallClient
 from trafficlight.internals.test import Test
-from trafficlight.tests.video.handle_invite_base import InviteLinksMixin
+from trafficlight.sample_tests.video.handle_invite_base import InviteLinksMixin
 
 
-class TwoRegisteredInviteLinksTest(Test, InviteLinksMixin):
+class TwoGuestsInviteLinksTest(Test, InviteLinksMixin):
     def __init__(self) -> None:
         super().__init__()
         self._client_under_test([ElementCall()], "alice")
         self._client_under_test([ElementCall()], "bob")
 
     async def run(self, alice: ElementCallClient, bob: ElementCallClient) -> None:
-        await alice.register()
-        await bob.register()
-        await alice.set_display_name()
-        await bob.set_display_name()
+        await alice.guest_user()
+        await bob.guest_user()
         await self._run_test(alice, bob)
