+++ trafficlight/sample_tests/video/three_user_spotlight.py
